+++ src/browser-lib/cookie-consent/types.ts
@@ -14,10 +14,16 @@ export const defaultConsentChoice = {
 };
 export type CookiePolicyName = typeof COOKIE_POLICY_NAMES[number];
 
-export const SERVICE_TYPES = ["posthog", "gleap", "bugsnag"] as const;
+export const SERVICE_TYPES = [
+  "posthog",
+  "gleap",
+  "bugsnag",
+  "hubspot",
+] as const;
 export type ServiceType = typeof SERVICE_TYPES[number];
 export const servicePolicyMap: Record<ServiceType, CookiePolicyName> = {
   posthog: "statistics",
   gleap: "statistics",
   bugsnag: "statistics",
+  hubspot: "statistics",
 };